server file renamed and tests fixed
