+++ server/swagger.yml
@@ -21,7 +21,7 @@ paths:
           description: "Server is running"
   /testdata:
     get:
-      operationId: "server.read_test_data"
+      operationId: "testdataserver.read_test_data"
       tags:
       - "Testdata"
       summary: "Get test data"
