+++ tests/test_server.py
@@ -1,6 +1,6 @@
 import pytest
 
-import server
+import testdataserver as server
 
 
 @pytest.fixture(scope='function')
